requirements.txt
absl-py==2.0.0
altair==5.2.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
attrs==23.1.0
black==23.11.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cloudpickle==3.0.0
cmake==3.27.7
comm==0.2.0
contourpy==1.2.0
customtkinter==5.2.1
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.0
decorator==5.1.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@017abbfa5f2c2a2afa045200c2af9ccf2fc6227f
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fonttools==4.45.1
fpdf==1.7.2
ftfy==6.1.3
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.3
hydra-core==1.3.2
idna==3.6
importlib-metadata==6.8.0
iopath==0.1.9
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lit==17.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.3
parso==0.8.3
pathspec==0.11.2
pdfkit==1.0.0
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.0.0
plotly==5.18.0
portalocker==2.8.2
prompt-toolkit==3.0.41
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybase64==1.3.1
pycocotools==2.0.7
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
Python-IO==0.3
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.28.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ultralytics==8.0.227
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.12
Werkzeug==3.0.1
yacs==0.1.8
zipp==3.17.0
